requirements.txt
# Pinned setuptools to avoid deprecation warning:
setuptools<80

# Core
pandas
pyyaml
openpyxl
odfpy
simplejson
humanize
dotenv

# Web
Flask
Flask-Admin==2.0.0a4
Flask-Assets
Flask-WTF
Flask-SQLAlchemy
cssmin
plotly
celery[redis]
gunicorn
requests
watchdog
markdown-it-py
beautifulsoup4

# Database
sqlalchemy
SQLAlchemy-Utc
pymysql
cryptography

# Testing
pytest
pytest-env
pytest-cov
freezegun

# Streamlit app:
# matplotlib
# mysql-connector-python
# prettytable
